Split existing code into modules
Add modules: configs, locking, outputs, runners, validators, misc and globalvars
Remove ansible_deployer/command_line.py
Adjust setup and increase version
