+++ source/modules/globalvars.py
@@ -1,0 +1,5 @@
+"""Global variables"""
+
+APP_CONF = "/etc/ansible-deployer"
+CFG_PERMISSIONS = "0o644"
+SUBCOMMANDS = ("run", "list", "lock", "unlock", "verify", "show")
